# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Intela API
  version: "0.1.0"
  description: |
    Reconocimiento de obras y distribucion de ingresos por propiedad
    intelectual para REDES SGC.

    Toda cifra que el sistema produzca debe ser explicable hasta su origen
    (fuente, reporte, regla aplicada). El reglamento lo exige y la auditoria
    lo revisa.
  license:
    name: Propietario
    url: https://redescritores.com
servers:
  # El host va como variable y no como literal a proposito: Amplify lo genera
  # al crear la rama, asi que no se conoce hasta el primer `terraform apply`.
  # Lo imprime `terraform output -raw url` (ADR 0014). Fijarlo aqui obligaria a
  # tocar el contrato cada vez que el entorno se recrea.
  - url: https://{host}/api
    description: Produccion, detras de Amplify Hosting
    variables:
      host:
        default: main.intela.amplifyapp.com
        description: Dominio de la rama en Amplify, o el dominio propio si se configura uno.
  - url: http://localhost/api
    description: Desarrollo, detras de nginx
  - url: http://localhost:8080
    description: Desarrollo, API directa
...
paths:
...
  /auth/session:
    post:
      tags: [sesion]
      summary: Iniciar sesion
      description: |
        Cambia unas credenciales por un token de sesion.

        Es la unica operacion que se llama sin token, de ahi el `security: []`.

        Devuelve el mismo 401 tanto si el correo no esta en el padron como si
        la clave no cuadra, y con el mismo mensaje: distinguirlos convertiria
        el login en una forma de averiguar que correos estan dados de alta.
      operationId: iniciarSesion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credenciales"
            example:
              email: ana@redes.co
              clave: la-clave-de-ana
      responses:
        "200":
          description: Sesion abierta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sesion"
              example:
                token: 9qYQ2vJmXk3pR7wLfNbTzA5cHdEgSuVxYnMoPiKjRlQ
                expira: "2026-08-31T00:00:00Z"
                usuario:
                  id: usr-ana
                  email: ana@redes.co
                  nombre: Ana Escritora
                  rol: titular
                  titular_id: tit-ana
        "400":
          description: Falta el correo o la clave, o el cuerpo no es JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: email y clave son obligatorios
        "401":
          description: Credenciales invalidas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: credenciales invalidas

    get:
      tags: [sesion]
      summary: Quien soy
      description: |
        Devuelve el usuario dueno del token presentado.

        Sirve para que un cliente pinte la sesion sin guardar el usuario en el
        navegador, donde no se puede confiar en el.
      operationId: sesionActual
      responses:
        "200":
          description: El usuario de la sesion.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
              example:
                id: usr-ana
                email: ana@redes.co
                nombre: Ana Escritora
                rol: titular
                titular_id: tit-ana
        "401":
          description: Falta el token, o esta caducado o revocado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: sesion invalida o expirada

    delete:
      tags: [sesion]
      summary: Cerrar sesion
      description: |
        Revoca el token presentado.

        Es idempotente: cerrar una sesion que ya estaba cerrada responde igual.
        Decir "ese token no existia" contaria algo sobre tokens ajenos.
      operationId: cerrarSesion
      responses:
        "204":
          description: Sesion cerrada. Sin cuerpo.
        "401":
          description: Falta el token, o esta caducado o revocado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: sesion invalida o expirada
  /afiliaciones:
    post:
      tags: [afiliacion]
      summary: Solicitar afiliacion
      description: |
        Recibe el alta de un titular nuevo y la deja en estado `pendiente`.

        Va sin sesion: quien se da de alta todavia no es afiliado.

        El RUT actualizado y la certificacion bancaria son obligatorios
        (`R-12` / `RD 13.1.6`). El IPI se puede omitir en esta peticion;
        una persona natural no se admite al padron sin el (`RD 3`). Hay
        `PATCH /afiliaciones/{id}/ipi` para completarlo despues.

        La clave (8 a 72 caracteres) se recoge aqui: al admitir se crea
        la cuenta con la que el titular entra.

        Este endpoint es publico y esta limitado por IP. Un correo ya
        afiliado no responde 409: enumeraria el padron.

        Si el aspirante declara pertenecer a otra sociedad de gestion
        colectiva del mismo genero y no adjunta la renuncia, responde 409
        con la explicacion de `R-28`.
      operationId: solicitarAfiliacion
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SolicitudAfiliacion"
      responses:
        "201":
          description: Solicitud creada, pendiente de admision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Afiliacion"
              example:
                id: afil-9qYQ2vJmXk3pR7wLfNbTzA5cHdEgSuVxYnMoPiKjRlQ
                nombre: Ana Escritora
                email: ana@redes.co
                documento_identidad: "12345678"
                ipi: IPI-00000001
                subtipo: socio
                estado: pendiente
                elegible_anticipo: false
                tiene_rut: true
                tiene_certificacion_bancaria: true
                tiene_renuncia: false
                titular_id: ""
        "400":
          description: |
            Faltan datos o documentos, el adjunto no es un PDF o una imagen,
            o no se pudo registrar (incluido un correo ya activo: no se
            distingue a proposito).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "r-12: el rut actualizado y la certificacion bancaria son obligatorios"
        "409":
          description: Conflicto de exclusividad (`R-28`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "r-28: no se acepta como afiliado a quien pertenezca a otra sociedad de gestion colectiva del mismo genero sin renuncia previa y expresa"
        "413":
          description: El cuerpo supera 18 MiB (tres documentos de 5 MiB mas el formulario).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el cuerpo supera el tamano maximo permitido
        "429":
          description: Demasiadas solicitudes desde la misma IP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: demasiadas solicitudes, reintente en un momento
        "503":
          description: |
            Esta instalacion no cablea la admision. Es el caso de `cmd/lambda`:
            el sistema de ficheros no sirve de boveda para el RUT y la
            certificacion bancaria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el alta de afiliacion no esta configurada en esta instalacion
  /afiliaciones/{id}/ipi:
    patch:
      tags: [afiliacion]
      summary: Completar IPI de una solicitud pendiente
      description: |
        Rellena el IPI que el alta permite omitir. Va sin sesion: el
        `{id}` es un token opaco de 256 bits. Una persona natural no se
        admite al padron sin este identificador (`RD 3`).
      operationId: completarIPI
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la solicitud, opaco.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ipi]
              properties:
                ipi:
                  type: string
                  description: Identificador IPI de la persona.
            example:
              ipi: IPI-00000001
      responses:
        "200":
          description: IPI guardado. La solicitud sigue pendiente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Afiliacion"
        "400":
          description: IPI vacio, o no se pudo completar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el ipi es obligatorio para admitir a una persona natural
        "404":
          description: No hay una solicitud con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: solicitud no encontrada
        "409":
          description: La solicitud ya no esta pendiente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: la solicitud no esta pendiente de admision
        "429":
          description: Demasiadas solicitudes desde la misma IP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: demasiadas solicitudes, reintente en un momento
        "503":
          description: |
            Esta instalacion no cablea la admision. Es el caso de `cmd/lambda`:
            el sistema de ficheros no sirve de boveda para el RUT y la
            certificacion bancaria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el alta de afiliacion no esta configurada en esta instalacion
  /afiliaciones/{id}/aprobar:
    post:
      tags: [afiliacion]
      summary: Aprobar afiliacion
      description: |
        Admite una solicitud pendiente, crea la fila del padron y la
        cuenta en `usuarios` con la clave que el aspirante eligio al
        solicitar.

        Solo el rol `administrador`, que en este andamiaje representa al
        Consejo Directivo (`RS 5.2`). A partir de la admision el subtipo
        gobierna el derecho a pedir anticipo: solo un Socio lo tiene (`R-30`).
      operationId: aprobarAfiliacion
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la solicitud, opaco.
          schema:
            type: string
      responses:
        "200":
          description: Solicitud admitida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Afiliacion"
              example:
                id: afil-9qYQ2vJmXk3pR7wLfNbTzA5cHdEgSuVxYnMoPiKjRlQ
                nombre: Ana Escritora
                email: ana@redes.co
                documento_identidad: "12345678"
                ipi: IPI-00000001
                subtipo: socio
                estado: admitido
                elegible_anticipo: true
                tiene_rut: true
                tiene_certificacion_bancaria: true
                tiene_renuncia: false
                titular_id: tit-9qYQ2vJmXk3pR7wLfNbTzA5cHdEgSuVxYnMoPiKjRlQ
        "400":
          description: Falta el IPI para admitir a una persona natural.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el ipi es obligatorio para admitir a una persona natural
        "401":
          description: Falta el token, o esta caducado o revocado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: sesion invalida o expirada
        "403":
          description: El rol de la sesion no es el del Consejo Directivo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: no autorizado
        "404":
          description: No hay una solicitud con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: solicitud no encontrada
        "409":
          description: La solicitud ya estaba resuelta, o el padron ya tiene esa fila.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: la solicitud no esta pendiente de admision
        "503":
          description: |
            Esta instalacion no cablea la admision. Es el caso de `cmd/lambda`:
            el sistema de ficheros no sirve de boveda para el RUT y la
            certificacion bancaria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el alta de afiliacion no esta configurada en esta instalacion
  /afiliaciones/{id}/rechazar:
    post:
      tags: [afiliacion]
      summary: Rechazar afiliacion
      description: |
        Cierra una solicitud pendiente. Libera el correo y el IPI para
        que el aspirante pueda volver a presentarse. Solo el Consejo
        Directivo (`RS 5.2`).
      operationId: rechazarAfiliacion
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la solicitud, opaco.
          schema:
            type: string
      responses:
        "200":
          description: Solicitud rechazada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Afiliacion"
              example:
                id: afil-9qYQ2vJmXk3pR7wLfNbTzA5cHdEgSuVxYnMoPiKjRlQ
                nombre: Ana Escritora
                email: ana@redes.co
                documento_identidad: "12345678"
                ipi: IPI-00000001
                subtipo: socio
                estado: rechazado
                elegible_anticipo: false
                tiene_rut: true
                tiene_certificacion_bancaria: true
                tiene_renuncia: false
                titular_id: ""
        "401":
          description: Falta el token, o esta caducado o revocado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: sesion invalida o expirada
        "403":
          description: El rol de la sesion no es el del Consejo Directivo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: no autorizado
        "404":
          description: No hay una solicitud con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: solicitud no encontrada
        "409":
          description: La solicitud ya estaba resuelta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: la solicitud no esta pendiente de admision
        "503":
          description: |
            Esta instalacion no cablea la admision. Es el caso de `cmd/lambda`:
            el sistema de ficheros no sirve de boveda para el RUT y la
            certificacion bancaria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: el alta de afiliacion no esta configurada en esta instalacion
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Token de sesion. Las sesiones expiran: el token no es una credencial
        permanente.

        La autenticacion no basta: cada grupo de rutas exige un rol del
        reglamento (`x-required-roles` en la operacion). Sin sesion la
        respuesta es 401; con sesion de un rol no permitido, 403. La
        matriz Rol -> capacidad esta en docs/architecture/roles.md.
...
  schemas:
...
    Credenciales:
      type: object
      required: [email, clave]
      properties:
        email:
          type: string
          format: email
          description: Correo con el que el usuario esta dado de alta.
        clave:
          type: string
          format: password
          description: |
            La contrasena en claro. Solo viaja en esta peticion y nunca se
            almacena: lo que se guarda es su hash bcrypt.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            Que fallo, en lenguaje de quien llama. Las respuestas de error van
            con `Content-Type: application/json`, no `text/plain`.
...
security:
  - bearerAuth: []